requirements.txt
git+https://github.com/AnswerDotAI/fasthtml.git
httpx
openai-whisper 
torch